requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
black==21.9b0
boto3==1.21.27
botocore==1.24.27
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
charset-normalizer==2.0.4
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==3.2.6
django-allauth==0.45.0
django-anymail==8.4
django-cacheops==6.0
django-cors-headers==3.7.0
django-cron==0.6.0
django-dbbackup==3.3.0
django-environ==0.8.1
django-extensions==3.1.3
django-galleryfield==1.3.dev0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-segno-qr==1.0.2
django-slugger==1.1.2
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
filelock==3.3.1
funcy==1.17
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
identify==2.3.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==1.0.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.2
pagacollect==1.0.5
path==16.2.0
path.py==12.5.0
pathspec==0.9.0
pause==0.3
phonenumberslite==8.12.28
Pillow==9.2.0
platformdirs==2.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.2
qrcode-artistic==2.1.0
redis==4.1.4
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
segno==1.3.3
simplejson==3.17.5
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.9.0
whitenoise==5.3.0
wrapt==1.14.0
